requirements.txt
kafka-python
pysolar
requests
pyspark
shapely
streamlit
streamlit-folium
pandas
folium
watchdog
geopy
haversine
pymongo